Standardize nodejs & python checks
This should fix the Node OOM error.
