+++ pkg/codegen/python/gen_test.go
@@ -16,7 +16,6 @@ package python
 
 import (
 	"fmt"
-	filesystem "io/fs"
 	"os"
 	"os/exec"
 	"path/filepath"
@@ -148,32 +147,6 @@ func buildVirtualEnv() error {
 	}
 
 	return nil
-}
-
-// Checks generated code for syntax errors with `python -m compile`.
-func pyCompileCheck(t *testing.T, codeDir string) {
-	pythonFiles := []string{}
-	err := filepath.Walk(codeDir, func(path string, info filesystem.FileInfo, err error) error {
-		require.NoError(t, err) // an error in the walk
-
-		if info.Mode().IsDir() && info.Name() == "venv" {
-			return filepath.SkipDir
-		}
-
-		if info.Mode().IsRegular() && strings.HasSuffix(info.Name(), ".py") {
-			path, err = filepath.Abs(path)
-			require.NoError(t, err)
-
-			pythonFiles = append(pythonFiles, path)
-		}
-		return nil
-	})
-	require.NoError(t, err)
-
-	ex, _, err := python.CommandPath()
-	require.NoError(t, err)
-	args := append([]string{"-m", "py_compile"}, pythonFiles...)
-	test.RunCommand(t, "python syntax check", codeDir, ex, args...)
 }
 
 func pyTestCheck(t *testing.T, codeDir string) {
