+++ pkg/codegen/testing/test/testdata/kubernetes-pod-pp/nodejs/kubernetes-pod.ts
@@ -15,7 +15,7 @@ const bar = new kubernetes.core.v1.Pod("bar", {
             resources: {
                 limits: {
                     memory: "20Mi",
-                    cpu: 0.2,
+                    cpu: "0.2",
                 },
             },
         }],
